[Add] model folder with User schema
