+++ README.md
@@ -44,3 +44,14 @@ _All this and more with MyVote!_
 ### Tech Stack
 
 ![MyVote Tech_stack](assets/myVote_techstack.jpg)
+
+
+### MongoDB Atlas setup
+
+- Create new Cluster named myvotecluster
+- Create a new Database user
+  - under Database Access, add new database user
+  - give user read adn write any database access
+- Add IP access
+  - under Network Access, add ip address
+  - add your current ip address
