+++ db/index.js
@@ -18,7 +18,4 @@ const connectDB = async () => {
   }
 };
 
-
-
-
 module.exports = connectDB;
